django_env/requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.12.4
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
